Moved Shop Class to protocols
